chore: move factories to __tests__
