+++ __tests__/utils/calldata/parser/parser-0-1.1.0.test.ts
@@ -1,5 +1,5 @@
 import { AbiParser1 } from '../../../../src/utils/calldata/parser/parser-0-1.1.0';
-import { getFunctionAbi, getInterfaceAbi } from '../../../../__mocks__/factories/abi';
+import { getFunctionAbi, getInterfaceAbi } from '../../../factories/abi';
 
 describe('AbiParser1', () => {
   test('should create an instance', () => {
