+++ __tests__/utils/calldata/parser/parser.test.ts
@@ -5,7 +5,7 @@ import {
 } from '../../../../src/utils/calldata/parser';
 import { AbiParser2 } from '../../../../src/utils/calldata/parser/parser-2.0.0';
 import { AbiParser1 } from '../../../../src/utils/calldata/parser/parser-0-1.1.0';
-import { getFunctionAbi, getInterfaceAbi } from '../../../../__mocks__/factories/abi';
+import { getFunctionAbi, getInterfaceAbi } from '../../../factories/abi';
 
 describe('createAbiParser', () => {
   test('should create an AbiParser2 instance', () => {
